refactor: refactored redundant test codes
